When table has been populated invoke onrendercomplete (#58)
* When table has been populated invoke onrendercomplete

So that the listener can be notified asynchronously that all the dependent data has been loaded

* raise a custom “ca-list-render-complete” event

to inform all listeners that the list has been rendered. Pass the loaded data back in the event.detail.data property in case the listener wants to do anything with it

* Don’t pass the data back as it might be large

Only want to inform the listeners that the list has been rendered. The listener can get hold of the data directly from the control if they need it